requirements.txt
Django>=5.0
python-decouple
dj-database-url
psycopg2-binary
gunicorn
whitenoise
python-dotenv
openai
supabase
python-dateutil
markdown
